Make diffusion component exports faithful to their schedulers
Sampler schedules are not integral. EulerDiscreteScheduler with five steps
over the SD1.x training schedule produces timesteps 999, 749.25, 499.5,
249.75, 0, and the same is true of every scheduler that interpolates the
training grid. Declaring the denoiser's timestep port as INT64 silently
truncated those values, so an exported UNet could not reproduce its own
pipeline no matter how accurate its weights were. Declare the port as FLOAT
in DenoisingTask and ControlNetTask, which is also what diffusers feeds its
UNet.

Shape inference gives the denoiser estimate and the VAE decoder sample
anonymous symbols (_d0, _d1, _d2) because the spatial relationship is a
property of the configured block stack, not of the graph. A consumer wiring
these components into a loop cannot tell that the estimate has the latent's
shape or that the decoded image scales with it. Republish named dimensions
on both. Denoisers that learn the variance emit twice the latent channels,
so derive the count from out_channels when the config has one and from
in_channels * (2 if learn_sigma else 1) otherwise, which keeps HunyuanDiT
and other learn_sigma configs correct without naming any model.

(cherry picked from commit df045d6a7cddbfd09e11541c059dbcf3a8b0b4e4)

+++ src/mobius/models/unet_parity_test.py
@@ -129,7 +129,7 @@ def test_unet_matches_diffusers():
         model,
         {
             "sample": sample.numpy(),
-            "timestep": timestep.numpy().astype(np.int64),
+            "timestep": timestep.numpy().astype(np.float32),
             "encoder_hidden_states": encoder_hidden_states.numpy(),
         },
     )
@@ -194,7 +194,7 @@ def test_unet_sd1x_mixed_block_types_matches_diffusers():
         model,
         {
             "sample": sample.numpy(),
-            "timestep": timestep.numpy().astype(np.int64),
+            "timestep": timestep.numpy().astype(np.float32),
             "encoder_hidden_states": encoder_hidden_states.numpy(),
         },
     )
@@ -273,7 +273,7 @@ def test_unet_lora_gate_parity():
 
     feed = {
         "sample": sample.numpy(),
-        "timestep": timestep.numpy().astype(np.int64),
+        "timestep": timestep.numpy().astype(np.float32),
         "encoder_hidden_states": encoder_hidden_states.numpy(),
     }
     off, on = _run_onnx(
